Refactor auth utils and update user retrieval across routes
