+++ app/onboarding/page.tsx
@@ -4,7 +4,7 @@ import { useState, useEffect, useRef } from "react"
 import { useRouter } from "next/navigation"
 import { useAuthV2 as useAuth } from "@/hooks/use-auth-v2"
 import { supabase } from "@/lib/supabase"
-import { useSafeRedirect } from "@/lib/auth-utils"
+import { useSafeRedirect } from '@/lib/client-utils'
 import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Button } from "@/components/ui/button"
 import { Input } from "@/components/ui/input"
